refactored form and added new useform custom hook
